reflect review comments
