+++ lib/generators/audited/templates/add_sql_to_audits.rb
@@ -1,10 +1,9 @@
 class <%= migration_class_name %> < <%= migration_parent %>
-    def self.up
-      add_column :audits, :sql, :text
-    end
-  
-    def self.down
-      remove_column :audits, :sql
-    end
-  end
-  
+	def self.up
+		add_column :audits, :sql, :text
+	end
+
+	def self.down
+		remove_column :audits, :sql
+	end
+end
